Add interface to get store weight from memory
Some of the available glance stores like file, cinder etc has
capability to reuse already initiated driver
(DRIVER_REUSABLE = 0b01000000). In Caracal we have added a
feature to sort image locations based on store weight. As RBD
driver of glance does not have this reuse capability, during
image list API call it initializes the RBD driver for each of
the available image which is causing noticable delay in list
call.

To avoid this, introducing new interface in glance_store which
will directly get the weight of the store from memory and return
it back to user.

NOTE: Since the current module does not have any test coverage,
corresponding tests will be added in glance once we release
glance_store with this change.

Related-Bug: #2086675
Change-Id: If6861f3271f50680af271090b516f2add6e392b9