Followup: Add try/finally block
This is a followup of 55cefedd169b84d713a8ff080d909483ea51c913
where we fix the file pointer issue with glance cinder powerflex
driver.
It was suggested to use try/finally block to restore file pointer
if a call to the _get_device_size method is made which is done
in this patch.

Change-Id: I0da7815e8b928dbfc7b1d03c2558709214592c9d