Disable do_secure_hash for glance_store-src-ceph-tempest
Few of the tests are either failing while deleting the resource
due to hash calculation of sanpshot or backup is in progress or
timing out for waiting the same. I think if we want to enable
hash calculation for snapshot or backup then we should implement
wait logic there to complete hash calculation before deleting
the resource. This can be implemented later if required.

Change-Id: I843f33925c54bcc0cbaa1dcd1e9918d72e9a6bc4