Run ceph-tempest job specifically for RBD
Previously, the glance_store-src-ceph-tempest job executed across
all backends, which was unnecessary. I have now disabled it for
backends other than RBD, so it will now run exclusively for changes
related to the RBD backend.

Additionally, it has been configured to not run when changes are made
in unit or functional tests.

Change-Id: I340fbaf3d9eb8802e70cf84242bda5dc7da90677